some changes

+++ src/main/java/com/geekster/expensestracker/model/AuthToken.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 public class AuthToken {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long tokenId;
+    private Integer tokenId;
 
     private String token;
